requirements.txt
flask
requests
psutil
pyyaml
waitress
